@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'rspec'
 require 'list_master'
 
@@ -6,14 +8,13 @@ require 'fixtures/active_record_fixtures'
 require 'fixtures/list_master_fixtures'
 
 # Configure ListMaster
-ListMaster.redis = Redis.connect :db => 9
+ListMaster.redis = Redis.new(db: 9)
 
 # Get fresh db on every test
 RSpec.configure do |config|
-
   config.include Support
 
-  config.before(:each) do
+  config.before do
     ListMaster.redis.redis.flushdb
   end
 
