@@ -1,6 +1,8 @@
+# frozen_string_literal: true
+
+# rubocop:disable Style/TrivialAccessors
 module ListMaster
   class Dsl
-
     def initialize
       @scope = :unscoped
       @sets = []
@@ -8,53 +10,50 @@ module ListMaster
       @remove_sets = true
     end
 
-    def namespace namespace
+    def namespace(namespace)
       @namespace = namespace
     end
 
-    def remove_sets remove
+    def remove_sets(remove)
       @remove_sets = remove
     end
 
     #
     # Associating this list master with a model
     #
-    def model model_class
+    def model(model_class)
       @model = model_class
     end
-
 
     #
     # Specify a scope to query for when updating redis sets
     #
-    def scope scope_name
+    def scope(scope_name)
       @scope = scope_name
     end
-
 
     #
     # Add an association to include as part of the process query
     #
-    def associated association
+    def associated(association)
       @associations << association
     end
-
 
     #
     # Defining sets to maintain
     #
-    def set *args
+    def set(*args)
       options = args.extract_options!
       @sets << {
         name: args.first.to_s,
         attribute: nil,
         descending: nil,
-        on: lambda {|m| m},
-        if: lambda {|m| true},
+        on: ->(m) { m },
+        if: ->(_m) { true },
         multi: nil,
         single: false,
       }.merge(options)
     end
-
   end
 end
+# rubocop:enable Style/TrivialAccessors
