@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module ListMaster
-  VERSION = '0.3'
+  VERSION = '1.0.0'
 end
